Updated script generating Linux-specific Steam depot to reflect a cloned project lacking any changed files that deviate from the ones tracked by Git, pull is performed before generating

+++ omni_gen_steam_depot_linux.sh
@@ -1,8 +1,15 @@
 #!/bin/bash
 
+if [ $# -ne 1 ]; then
+	echo "Error: Wrong number of arguments"
+
+	exit 2
+fi
+
 LIBDIR=build
 
 rm build_linux.zip
 
+git pull $1
 make
 zip -r build_linux.zip omni $LIBDIR/*.so res lvl player/cfg player/doc omni.py catch_out.py lvl_spec.json stat.json installscript.vdf
